+++ lib/rom/sql/extensions/postgres.rb
@@ -1,0 +1,3 @@
+require 'rom/sql/extensions/postgres/commands'
+require 'rom/sql/extensions/postgres/types'
+require 'rom/sql/extensions/postgres/inferrer'