+++ lib/rom/sql/schema/inferrer.rb
@@ -1,27 +1,43 @@
 module ROM
   module SQL
     class Schema < ROM::Schema
+      # @api private
       class Inferrer
         extend ClassMacros
 
-        defines :type_mapping, :pk_type
+        defines :ruby_type_mapping, :numeric_pk_type, :db_type, :db_registry
 
-        type_mapping(
-          integer: Types::Strict::Int,
-          string: Types::Strict::String,
-          date: Types::Strict::Date,
-          datetime: Types::Strict::Time,
-          boolean: Types::Strict::Bool,
-          decimal: Types::Strict::Decimal,
-          blob: Types::Strict::String
+        ruby_type_mapping(
+          integer: Types::Int,
+          string: Types::String,
+          date: Types::Date,
+          datetime: Types::Time,
+          boolean: Types::Bool,
+          decimal: Types::Decimal,
+          float: Types::Float,
+          blob: Types::Blob
         ).freeze
 
-        pk_type Types::Serial
+        numeric_pk_type Types::Serial
 
-        attr_reader :dsl
+        db_registry Hash.new(self)
 
-        def initialize(dsl)
-          @dsl = dsl
+        def self.inherited(klass)
+          super
+
+          Inferrer.db_registry[klass.db_type] = klass unless klass.name.nil?
+        end
+
+        def self.[](type)
+          Class.new(self) { db_type(type) }
+        end
+
+        def self.get(type)
+          if !db_registry.key?(type) && ROM::SQL.available_extension?(type)
+            ROM::SQL.load_extensions(type)
+          end
+
+          db_registry[type]
         end
 
         # @api private
@@ -29,31 +45,33 @@ module ROM
           columns = gateway.connection.schema(dataset)
           fks = fks_for(gateway, dataset)
 
-          columns.each do |(name, definition)|
-            dsl.attribute name, build_type(definition.merge(foreign_key: fks[name]))
+          columns.each_with_object({}) do |(name, definition), attrs|
+            type = build_type(definition.merge(foreign_key: fks[name]))
+            attrs[name] = type.meta(name: name)
           end
-
-          pks = columns
-            .map { |(name, definition)| name if definition.fetch(:primary_key) }
-            .compact
-
-          dsl.primary_key(*pks) if pks.any?
-
-          dsl.attributes
         end
 
         private
 
-        # @api private
-        def build_type(primary_key: , type: , allow_null: , foreign_key: , **rest)
+        def build_type(primary_key:, db_type:, type:, allow_null:, foreign_key:, **rest)
           if primary_key
-            self.class.pk_type
+            map_pk_type(type, db_type)
           else
-            type = self.class.type_mapping.fetch(type)
-            type = type.optional if allow_null
-            type = type.meta(foreign_key: true, relation: foreign_key) if foreign_key
-            type
+            mapped_type = map_type(type, db_type)
+            mapped_type = mapped_type.optional if allow_null
+            mapped_type = mapped_type.meta(foreign_key: true, relation: foreign_key) if foreign_key
+            mapped_type
           end
+        end
+
+        def map_pk_type(_ruby_type, _db_type)
+          self.class.numeric_pk_type.meta(primary_key: true)
+        end
+
+        def map_type(ruby_type, db_type)
+          self.class.ruby_type_mapping.fetch(ruby_type) {
+            raise UnknownDBTypeError, "Cannot find corresponding type for #{ruby_type || db_type}"
+          }
         end
 
         # @api private
@@ -65,7 +83,6 @@ module ROM
           end
         end
 
-        # @api private
         def build_fk(columns: , table: , **rest)
           if columns.size == 1
             [columns[0], table]