+++ lib/rom/sql/types.rb
@@ -5,7 +5,11 @@ module ROM
     module Types
       include ROM::Types
 
-      Serial = Strict::Int.constrained(gt: 0).meta(primary_key: true)
+      Serial = Int.constrained(gt: 0).meta(primary_key: true)
+
+      Blob = Dry::Types::Definition
+        .new(Sequel::SQL::Blob)
+        .constructor(Sequel::SQL::Blob.method(:new))
     end
   end
 end