+++ spec/unit/relation/inner_join_spec.rb
@@ -1,0 +1,22 @@
+RSpec.describe ROM::Relation, '#inner_join' do
+  subject(:relation) { container.relations.users }
+
+  include_context 'users and tasks'
+
+  with_adapters do
+    it 'joins relations using inner join' do
+      result = relation.inner_join(:tasks, user_id: :id).select(:name, :title)
+
+      expect(result.to_a).to eql([
+        { name: 'Jane', title: "Jane's task" },
+        { name: 'Joe', title: "Joe's task" }
+      ])
+    end
+
+    it 'raises error when column names are ambiguous' do
+      expect {
+        relation.inner_join(:tasks, user_id: :id).to_a
+      }.to raise_error(Sequel::DatabaseError, /ambiguous/)
+    end
+  end
+end