+++ spec/unit/relation/max_spec.rb
@@ -1,0 +1,11 @@
+RSpec.describe ROM::Relation, '#max' do
+  subject(:relation) { container.relations.users }
+
+  include_context 'users and tasks'
+
+  with_adapters do
+    it 'delegates to dataset and return value' do
+      expect(relation.max(:id)).to eql(2)
+    end
+  end
+end