+++ spec/unit/relation/rename_spec.rb
@@ -1,0 +1,21 @@
+RSpec.describe ROM::Relation, '#rename' do
+  subject(:relation) { container.relations.users }
+
+  include_context 'users and tasks'
+
+  before do
+    conf.relation(:users) do
+      def sorted
+        order(:id)
+      end
+    end
+  end
+
+  with_adapters do
+    it 'projects the dataset using new column names' do
+      renamed = relation.sorted.rename(id: :user_id, name: :user_name)
+
+      expect(renamed.first).to eql(user_id: 1, user_name: 'Jane')
+    end
+  end
+end